Split paginate() and parse_money() into their own utilities.py module.
